* Begun coding for audio engine. This is _very_ preliminary.  * Fixed a warning on Opengl + mingw code.
git-svn-id: file:///home/michael/svn/yuni/trunk@849 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/window/msw/opengl.cpp
@@ -31,7 +31,7 @@ namespace Window
 			{
 				sizeof (PIXELFORMATDESCRIPTOR), 1,
 				PFD_DRAW_TO_WINDOW | PFD_SUPPORT_OPENGL | PFD_DOUBLEBUFFER,
-				PFD_TYPE_RGBA, pBitDepth, 0, 0, 0, 0, 0, 0,	// Color Bits Ignored
+				PFD_TYPE_RGBA, (BYTE)pBitDepth, 0, 0, 0, 0, 0, 0,	// Color Bits Ignored
 				// No Alpha Buffer
 				0,
 				// Shift Bit Ignored
